+++ src/floor_item.cpp
@@ -27,11 +27,7 @@
 
 #include "resources/image.h"
 
-FloorItem::FloorItem(unsigned int id,
-                     unsigned int itemId,
-                     unsigned short x,
-                     unsigned short y,
-                     Map *map):
+FloorItem::FloorItem(int id, int itemId, int x, int y, Map *map):
     mId(id),
     mX(x),
     mY(y),
@@ -52,7 +48,7 @@ FloorItem::~FloorItem()
     delete mItem;
 }
 
-unsigned int FloorItem::getItemId() const
+int FloorItem::getItemId() const
 {
     return mItem->getId();
 }
